move CometExecRule into separate file
